+++ qualang_tools/results/data_handler/data_processors.py
@@ -48,6 +48,7 @@ class DataProcessor(ABC):
 
 class MatplotlibPlotSaver(DataProcessor):
     file_format: str = "png"
+    nested_separator: str = "."
 
     def __init__(self, file_format=None):
         if file_format is not None:
@@ -65,11 +66,14 @@ class MatplotlibPlotSaver(DataProcessor):
         self.data_figures = {}
 
         for keys, val in iterate_nested_dict(data):
-            if isinstance(val, plt.Figure):
-                path = Path("/".join(keys)).with_suffix(self.file_suffix)
+            if not isinstance(val, plt.Figure):
+                continue
 
-                self.data_figures[path] = val
-                update_nested_dict(data, keys, f"./{path}")
+            file_end = Path(keys[-1]).with_suffix(self.file_suffix)
+            path = Path(self.nested_separator.join(keys[:-1] + [str(file_end)]))
+
+            self.data_figures[path] = val
+            update_nested_dict(data, keys, f"./{path}")
 
         return data
 
@@ -84,6 +88,7 @@ DEFAULT_DATA_PROCESSORS.append(MatplotlibPlotSaver)
 class NumpyArraySaver(DataProcessor):
     merge_arrays: bool = True
     merged_array_name: str = "arrays.npz"
+    nested_separator: str = "."
 
     def __init__(self, merge_arrays=None, merged_array_name=None):
         if merge_arrays is not None:
@@ -100,12 +105,12 @@ class NumpyArraySaver(DataProcessor):
             if not isinstance(val, np.ndarray):
                 continue
 
-            path = Path("/".join(keys))
+            path = Path(self.nested_separator.join(keys))
             self.data_arrays[path] = val
             if self.merge_arrays:
                 update_nested_dict(data, keys, f"./{self.merged_array_name}#{path}")
             else:
-                update_nested_dict(data, keys, f"./{path.with_suffix('.npy')}")
+                update_nested_dict(data, keys, f"./{path}.npy")
         return data
 
     def post_process(self, data_folder: Path):
@@ -115,7 +120,7 @@ class NumpyArraySaver(DataProcessor):
                 np.savez(data_folder / self.merged_array_name, **arrays)
         else:
             for path, arr in self.data_arrays.items():
-                np.save(data_folder / path.with_suffix(".npy"), arr)
+                np.save(data_folder / f"{path}.npy", arr)
 
 
 DEFAULT_DATA_PROCESSORS.append(NumpyArraySaver)
@@ -125,6 +130,7 @@ class XarraySaver(DataProcessor):
     merge_arrays: bool = False
     merged_array_name: str = "xarrays"
     file_format: str = "hdf5"
+    nested_separator: str = "."
 
     def __init__(self, merge_arrays=None, merged_array_name=None, file_format=None):
         if merge_arrays is not None:
@@ -150,13 +156,13 @@ class XarraySaver(DataProcessor):
             if not isinstance(val, xr.Dataset):
                 continue
 
-            path = Path("/".join(keys))
+            path = Path(self.nested_separator.join(keys))
             self.data_arrays[path] = val
             if self.merge_arrays:
                 merged_array_name = Path(self.merged_array_name).with_suffix(self.file_suffix)
                 update_nested_dict(data, keys, f"./{merged_array_name}#{path}")
             else:
-                update_nested_dict(data, keys, f"./{path.with_suffix(self.file_suffix)}")
+                update_nested_dict(data, keys, f"./{path}{self.file_suffix}")
         return data
 
     def save_merged_netcdf_arrays(self, path: Path, arrays: dict):
@@ -183,7 +189,7 @@ class XarraySaver(DataProcessor):
                     ) from e
         else:
             for path, array in self.data_arrays.items():
-                array.to_netcdf(data_folder / path.with_suffix(self.file_suffix))
+                array.to_netcdf(data_folder / f"{path}{self.file_suffix}")
 
 
 try: